DRAFT: Remove most of the busywaiting
DO NOT MERGE AS-IS - this is to be considered 32-bit x86 Linux-specific due to me struggling to make the headers behave properly and also needing an API exposed.

pd.exe tends to go down to about 9% of a CPU core in the main menu on My Machine(tm) at the default screen resolution.